+++ dev/integration_tests/ios_host_app/PodfileMissingPostInstall
@@ -10,7 +10,3 @@ end
 target 'FlutterUITests' do
   inherit! :search_paths
 end
-
-post_install do |installer|
-  flutter_post_install(installer)
-end
